Reorganise code so that it's all under src/. Remove unused old debug functions. Group MLX window code in src/window
